services/scraper/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
playwright>=1.49.0
redis>=5.2.0
asyncpg>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
httpx>=0.28.0
minio>=7.2.0
beautifulsoup4>=4.12.0
pillow>=11.0.0
